+++ R/read_experiment_statistics.R
@@ -1,6 +1,9 @@
 read_experiment_statistics <- function(file) {
   checkmate::assert_string(file)
-  checkmate::assert_file_exists(file, extension = "csv")
+  checkmate::assert_file_exists(file)
+
+  assert_behaviorspace_file(file)
+  assert_behaviorspace_file_output(file)
 
   out <-
     file |>
@@ -9,7 +12,8 @@ read_experiment_statistics <- function(file) {
       na = c("", "N/A"),
       skip = 6,
       progress = FALSE,
-      show_col_types = FALSE
+      show_col_types = FALSE,
+      skip_empty_rows = TRUE
     ) |>
     dplyr::as_tibble() |>
     janitor::clean_names()