+++ tests/testthat/test-read_experiment_lists.R
@@ -1,11 +1,11 @@
 testthat::test_that("`read_experiment_lists()` | General test", {
-  test_file <- tempfile(fileext = ".csv")
+  test_file <- tempfile()
 
   c(
     'BehaviorSpace results (NetLogo 7.0.3), "Lists version 2.0"',
     paste0(
-      '"/home/danielvartan/.opt/netlogo-7-0-3/',
-      'models/IABM Textbook/chapter 4/Wolf Sheep Simple 5.nlogox"'
+      '"/opt/NetLogo 7-0-3/models/',
+      'IABM Textbook/chapter 4/Wolf Sheep Simple 5.nlogox"'
     ),
     '"Wolf Sheep Simple Model Analysis"',
     '"01/05/2026 07:33:45:117 -0300"',
@@ -38,13 +38,13 @@ testthat::test_that("`read_experiment_lists()` | General test", {
 })
 
 testthat::test_that("`read_experiment_lists()` | Message test", {
-  test_file <- tempfile(fileext = ".csv")
+  test_file <- tempfile()
 
   c(
     'BehaviorSpace results (NetLogo 7.0.3), "Lists version 2.0"',
     paste0(
-      '"/home/danielvartan/.opt/netlogo-7-0-3/',
-      'models/IABM Textbook/chapter 4/Wolf Sheep Simple 5.nlogox"'
+      '"/opt/NetLogo 7-0-3/models/',
+      'IABM Textbook/chapter 4/Wolf Sheep Simple 5.nlogox"'
     ),
     '"Wolf Sheep Simple Model Analysis"',
     '"01/05/2026 07:33:45:117 -0300"',
@@ -71,36 +71,32 @@ testthat::test_that("`read_experiment_lists()` | Error test", {
   # checkmate::assert_string(file)
 
   read_experiment_lists(
-    file = 1,
-    tidy_output = TRUE
+    file = 1
   ) |>
     testthat::expect_error()
 
-  # checkmate::assert_file_exists(file, extension = "csv")
-
-  test_file <- tempfile(fileext = ".txt")
-  test_file |> file.create()
+  # checkmate::assert_file_exists()
 
   read_experiment_lists(
-    file = tempfile(),
-    tidy_output = TRUE
-  ) |>
-    testthat::expect_error()
-
-  read_experiment_lists(
-    file = test_file,
-    tidy_output = TRUE
+    file = tempfile()
   ) |>
     testthat::expect_error()
 
   # checkmate::assert_flag(tidy_output)
 
-  test_file <- tempfile(fileext = ".csv")
+  test_file <- tempfile()
   test_file |> file.create()
 
   read_experiment_lists(
     file = test_file,
     tidy_output = 1
+  ) |>
+    testthat::expect_error()
+
+  # checkmate::assert_tibble(data)
+
+  read_experiment_lists.tidy_data(
+    data = 1
   ) |>
     testthat::expect_error()
 })