+++ tests/testthat/test-read_experiment_metadata.R
@@ -1,11 +1,11 @@
 testthat::test_that("`read_experiment_metadata()` | General test", {
-  test_file <- tempfile(fileext = ".csv")
+  test_file <- tempfile()
 
   c(
     'BehaviorSpace results (NetLogo 7.0.3), "Table version 2.0"',
     paste0(
-      '"/home/danielvartan/.opt/netlogo-7-0-3/',
-      'models/IABM Textbook/chapter 4/Wolf Sheep Simple 5.nlogox"'
+      '"/opt/NetLogo 7-0-3/models/',
+      'IABM Textbook/chapter 4/Wolf Sheep Simple 5.nlogox"'
     ),
     '"Wolf Sheep Simple Model Analysis"',
     '"01/05/2026 06:37:48:683 -0300"',
@@ -27,18 +27,8 @@ testthat::test_that("`read_experiment_metadata()` | General test", {
   ) |>
     readr::write_lines(test_file)
 
-  read_experiment_metadata(
-    file = test_file,
-    output_version = FALSE
-  ) |>
-    checkmate::expect_list(len = 5)
-
-  # if (isTRUE(output_version)) {
-
-  read_experiment_metadata(
-    file = test_file,
-    output_version = TRUE
-  ) |>
+  test_file |>
+    read_experiment_metadata() |>
     checkmate::expect_list(len = 6)
 })
 
@@ -46,36 +36,14 @@ testthat::test_that("`read_experiment_metadata()` | Error test", {
   # checkmate::assert_string(file)
 
   read_experiment_metadata(
-    file = 1,
-    output_version = FALSE
+    file = 1
   ) |>
     testthat::expect_error()
 
-  # checkmate::assert_file_exists(file, extension = "csv")
-
-  test_file <- tempfile(fileext = ".txt")
-  test_file |> file.create()
+  # checkmate::assert_file_exists()
 
   read_experiment_metadata(
-    file = tempfile(),
-    output_version = FALSE
-  ) |>
-    testthat::expect_error()
-
-  read_experiment_metadata(
-    file = test_file,
-    output_version = FALSE
-  ) |>
-    testthat::expect_error()
-
-  # checkmate::assert_flag(output_version)
-
-  test_file <- tempfile(fileext = ".csv")
-  test_file |> file.create()
-
-  read_experiment_metadata(
-    file = test_file,
-    output_version = 1
+    file = tempfile()
   ) |>
     testthat::expect_error()
 })