+++ NEWS.md
@@ -21,6 +21,7 @@
 - `parse_netlogo_list()` now have better heuristics.
 - `parse_netlogo_color()` was introduced to parse NetLogo color strings into approximate hex color codes.
 - `get_netlogo_shape()` was introduced to retrieve NetLogo shape definitions from the [`LogoShapes`](https://github.com/danielvartan/logoshapes) project.
+- `read_experiment()` was introduced to read and tidy BehaviorSpace experiment output files into R.
 - New unit tests were implemented.
 - A [new vignette](https://danielvartan.github.io/logolink/articles/visualizing-the-netlogo-world.html) showing how to visualize the NetLogo world using [`ggplot2`](https://ggplot2.tidyverse.org/) was added.
 - The documentation was updated to reflect the changes in the package.